Key.Resolve: return a new Key object instead of internal mutation
The current Key.Resolve() API expects the Key object to mutate itself internally
to reflect the refreshed state. This is difficult to implement when a provider
splits different key types into several internal implementers, e.g. secretKey,
privateKey, publicKey, etc.

An attribute refresh could fundamentally change a key such that it is not
possible to perform the correct internal mutation as underlying concrete types
would have to change. For example, if a key ID previously yielded a secret key
and was represented by internal type secretKey, a refresh might have us discover
that the ID now refers to a public key. We cannot change the type without
returning an entirely new Key object.

Signed-off-by: Jonas Köhnen <jonas.koehnen@sap.com>